python-server/requirements.txt
grpcio
grpcio-tools
pandas
numpy
scipy
statsmodels
openpyxl